big ahh refactor
