+++ crons/DeleteExpiredCredentials.go
@@ -2,10 +2,10 @@ package crons
 
 import (
 	"github.com/robfig/cron/v3"
-	"mini-alt/storage"
+	"mini-alt/storage/db"
 )
 
-func SetupDeleteExpiredCredentials(store storage.Store) {
+func SetupDeleteExpiredCredentials(store *db.Store) {
 	c := cron.New()
 	_, err := c.AddFunc("0 2 * * *", func() { store.DeleteExpiredCredentials() })
 	if err != nil {
