+++ crons/crons.go
@@ -1,8 +1,10 @@
 package crons
 
-import "mini-alt/storage"
+import (
+	"mini-alt/storage/db"
+)
 
-func StartupCronJobs(store storage.Store) {
+func StartupCronJobs(store *db.Store) {
 	SetupDeleteExpiredCredentials(store)
 	SetupDeleteExpiredUsers(store)
 }
