+++ handlers/api/DeleteObject.go
@@ -2,7 +2,7 @@ package api
 
 import (
 	"github.com/gin-gonic/gin"
-	"mini-alt/storage"
+	"mini-alt/storage/disk"
 	"net/http"
 )
 
@@ -17,7 +17,7 @@ func (h *Handler) DeleteObject(c *gin.Context) {
 	// Also delete all files
 	err := h.Store.DeleteObject(bucketName, object)
 	if err == nil {
-		err := storage.DeleteObjectFile(bucketName, object)
+		err := disk.DeleteObjectFile(bucketName, object)
 		if err != nil {
 			println(err.Error())
 			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
