+++ storage/disk/GetObject.go
@@ -1,0 +1,11 @@
+package disk
+
+import "os"
+
+func GetObject(bucketName, objectKey string) (*os.File, error) {
+	path, err := GetObjectPath(bucketName, objectKey)
+	if err != nil {
+		return nil, err
+	}
+	return os.Open(path)
+}
