--- storage/sqlite.go
@@ -1,493 +1,0 @@
-package storage
-
-import (
-	"database/sql"
-	"errors"
-	"fmt"
-	"log"
-	"mini-alt/models"
-	"mini-alt/types"
-	"mini-alt/utils"
-	"time"
-)
-
-type SQLiteStore struct {
-	db *sql.DB
-}
-
-func NewSQLiteStore(dbPath string) (*SQLiteStore, error) {
-	db, err := sql.Open("sqlite3", dbPath)
-	if err != nil {
-		return nil, err
-	}
-
-	_, err = db.Exec(`PRAGMA foreign_keys = ON`)
-	if err != nil {
-		return nil, err
-	}
-
-	store := &SQLiteStore{db: db}
-	if err := store.initSchema(); err != nil {
-		return nil, err
-	}
-	if err := store.SeedInitialData(); err != nil {
-		return nil, err
-	}
-
-	return store, nil
-}
-
-func (s *SQLiteStore) initSchema() error {
-	schema := `
-	CREATE TABLE IF NOT EXISTS buckets (
-		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		name TEXT UNIQUE NOT NULL,
-		created_at DATETIME NOT NULL
-	);
-
-	CREATE TABLE IF NOT EXISTS objects (
-		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		bucket_name TEXT NOT NULL,
-		key TEXT NOT NULL,
-		size INTEGER,
-		last_modified DATETIME,
-		FOREIGN KEY(bucket_name) REFERENCES buckets(name) ON DELETE CASCADE,
-		UNIQUE(bucket_name, key)
-	);
-
-	CREATE INDEX IF NOT EXISTS idx_objects_bucket_name ON objects(bucket_name);
-
-	CREATE TABLE IF NOT EXISTS object_metadata (
-	    id INTEGER PRIMARY KEY AUTOINCREMENT,
-	    object_id INTEGER NOT NULL,
-	    acl TEXT,
-	    cache_control TEXT,
-	    content_disposition TEXT,
-	    content_encoding TEXT,
-	    content_language TEXT,
-	    content_length INTEGER,
-	    content_md5 TEXT,
-	    content_type TEXT,
-	    expires DATETIME,
-	    FOREIGN KEY(object_id) REFERENCES objects(id) ON DELETE CASCADE
-	);
-
-	CREATE TABLE IF NOT EXISTS  credentials (
-		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		access_key TEXT UNIQUE NOT NULL,
-		secret_key_encrypted TEXT NOT NULL,
-		user BOOLEAN NOT NULL,
-		expires_at DATE DEFAULT NULL,
-		status BOOLEAN NOT NULL DEFAULT TRUE,
-		name TEXT,
-		description TEXT,
-		created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-	);
-
-	CREATE TABLE IF NOT EXISTS users (
-	    id INTEGER PRIMARY KEY AUTOINCREMENT,
-	    username TEXT UNIQUE NOT NULL UNIQUE,
-	    password TEXT NOT NULL,
-	    token TEXT NOT NULL,
-		access_key TEXT NOT NULL,
-	    expires_at DATE DEFAULT NULL,
-	    created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-	)
-`
-	_, err := s.db.Exec(schema)
-	return err
-}
-
-func (s *SQLiteStore) SeedInitialData() error {
-	var count int
-	err := s.db.QueryRow(`SELECT COUNT(*) FROM users`).Scan(&count)
-	if err != nil {
-		return fmt.Errorf("failed to count users: %w", err)
-	}
-
-	if count > 0 {
-		return nil
-	}
-	accessKey, _, err := s.CreateCredentials("", "", "", true)
-	if err != nil {
-		return err
-	}
-
-	err = s.RegisterUser("admin", "admin", accessKey, "")
-
-	if err != nil {
-		return fmt.Errorf("failed to insert default user: %w", err)
-	}
-
-	return nil
-}
-
-func (s *SQLiteStore) PutObject(bucket, object string, size int64) (models.Object, error) {
-	now := time.Now()
-	row := s.db.QueryRow(`
-	INSERT INTO objects(bucket_name, key, size, last_modified)
-	VALUES (?, ?, ?, ?)
-	ON CONFLICT(bucket_name, key) DO UPDATE 
-	SET size = excluded.size, last_modified = excluded.last_modified
-	RETURNING id`,
-		bucket, object, size, now)
-
-	var id int64
-	if err := row.Scan(&id); err != nil {
-		return models.Object{}, err
-	}
-
-	return models.Object{Id: id, Key: object, Size: size, LastModified: now}, nil
-}
-
-func (s *SQLiteStore) PutBucket(bucket string) error {
-	now := time.Now()
-	_, err := s.db.Exec(`INSERT INTO buckets(name, created_at) VALUES (?, ?)`, bucket, now)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (s *SQLiteStore) PutMetadata(objectId int64, metadata types.Metadata) error {
-	_, err := s.db.Exec(`
-		INSERT INTO object_metadata (
-			object_id,
-			acl,
-			cache_control,
-			content_disposition,
-			content_encoding,
-			content_language,
-			content_length,
-			content_md5,
-			content_type,
-			expires
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		objectId,
-		metadata.ACL,
-		metadata.CacheControl,
-		metadata.ContentDisposition,
-		metadata.ContentEncoding,
-		metadata.ContentLanguage,
-		metadata.ContentLength,
-		metadata.ContentMD5,
-		metadata.ContentType,
-		metadata.Expires,
-	)
-	return err
-}
-
-func (s *SQLiteStore) ListObjects(bucket string) ([]models.Object, error) {
-	rows, err := s.db.Query(`SELECT key, size, last_modified FROM objects WHERE bucket_name = ?`, bucket)
-	if err != nil {
-		return nil, err
-	}
-	defer func(rows *sql.Rows) {
-		_ = rows.Close()
-	}(rows)
-
-	var objs []models.Object
-	for rows.Next() {
-		var o models.Object
-		if err := rows.Scan(&o.Key, &o.Size, &o.LastModified); err != nil {
-			return nil, err
-		}
-		objs = append(objs, o)
-	}
-	return objs, nil
-}
-
-func (s *SQLiteStore) ListBuckets() ([]models.Bucket, error) {
-	query := `
-		SELECT 
-			b.id,
-			b.name,
-			b.created_at,
-			IFNULL(COUNT(o.id), 0) as number_objects,
-			IFNULL(SUM(o.size), 0) as total_size
-		FROM 
-			buckets b
-		LEFT JOIN 
-			objects o ON o.bucket_name = b.name
-		GROUP BY 
-			b.id, b.name, b.created_at
-	`
-
-	rows, err := s.db.Query(query)
-	if err != nil {
-		return nil, err
-	}
-	defer func(rows *sql.Rows) {
-		_ = rows.Close()
-	}(rows)
-
-	var buckets []models.Bucket
-	for rows.Next() {
-		var b models.Bucket
-		if err := rows.Scan(&b.Id, &b.Name, &b.CreatedAt, &b.NumberObjects, &b.Size); err != nil {
-			println(err.Error())
-			return nil, err
-		}
-		buckets = append(buckets, b)
-	}
-	return buckets, nil
-}
-
-func (s *SQLiteStore) GetObject(bucket, key string) (models.Object, error) {
-	row := s.db.QueryRow(`
-		SELECT key, size, last_modified FROM objects WHERE bucket_name = ? AND key = ?`,
-		bucket, key)
-
-	var obj models.Object
-	obj.Key = key
-	if err := row.Scan(&obj.Key, &obj.Size, &obj.LastModified); err != nil {
-		return models.Object{}, errors.New("object not found")
-	}
-	return obj, nil
-}
-
-func (s *SQLiteStore) GetBucket(bucket string) (models.Bucket, error) {
-	row := s.db.QueryRow(`SELECT * FROM buckets WHERE name = ?`, bucket)
-	var b models.Bucket
-	b.Name = bucket
-	if err := row.Scan(&b.Id, &b.Name, &b.CreatedAt); err != nil {
-		return models.Bucket{}, err
-	}
-
-	return b, nil
-}
-
-func (s *SQLiteStore) DeleteObject(bucket, key string) error {
-	prefix := key + "%"
-	_, err := s.db.Exec(`DELETE FROM objects WHERE bucket_name = ? AND key LIKE ? ESCAPE '\'`, bucket, prefix)
-	return err
-}
-
-func (s *SQLiteStore) DeleteBucket(bucket string) error {
-	_, err := s.db.Exec(`DELETE FROM buckets WHERE name = ?`, bucket)
-	return err
-}
-
-func (s *SQLiteStore) CreateCredentials(name, description, expiresAt string, user bool) (accessKey, secretKey string, err error) {
-	accessKey = utils.GenerateRandomKey(16)
-	secretKey = utils.GenerateRandomKey(32)
-
-	encryptedSecret, err := utils.Encrypt(secretKey)
-	if err != nil {
-		return "", "", err
-	}
-
-	var expiresAtValue interface{}
-	if expiresAt == "" {
-		expiresAtValue = nil
-	} else {
-		expiresAtValue = expiresAt
-	}
-
-	_, err = s.db.Exec(`INSERT INTO credentials (access_key, secret_key_encrypted, user, expires_at, name, description) VALUES (?, ?, ?, ?, ?, ?)`, accessKey, encryptedSecret, user, expiresAtValue, name, description)
-	if err != nil {
-		return "", "", err
-	}
-
-	return accessKey, secretKey, nil
-}
-
-func (s *SQLiteStore) EditCredentials(accessKey string, name, description, expiresAt string, status bool) error {
-	var expiresAtValue interface{}
-	if expiresAt == "" {
-		expiresAtValue = nil
-	} else {
-		expiresAtValue = expiresAt
-	}
-
-	_, err := s.db.Exec(`UPDATE credentials SET expires_at = ?, status = ?, name = ?, description = ? WHERE access_key = ?`, expiresAtValue, status, name, description, accessKey)
-	return err
-}
-
-func (s *SQLiteStore) GetSecretKey(accessKey string) (string, error) {
-	row := s.db.QueryRow(`SELECT secret_key_encrypted FROM credentials WHERE access_key = ?`, accessKey)
-	var encrypted string
-	if err := row.Scan(&encrypted); err != nil {
-		return "", err
-	}
-
-	return utils.Decrypt(encrypted)
-}
-
-func (s *SQLiteStore) ListCredentials() ([]models.Credentials, error) {
-	rows, err := s.db.Query(`SELECT access_key, expires_at, created_at, status, name, description FROM credentials WHERE user = FALSE`)
-	if err != nil {
-		return nil, err
-	}
-	defer func(rows *sql.Rows) {
-		_ = rows.Close()
-	}(rows)
-
-	var credentials []models.Credentials
-	for rows.Next() {
-		var c models.Credentials
-		if err := rows.Scan(&c.AccessKey, &c.ExpiresAt, &c.CreatedAt, &c.Status, &c.Name, &c.Description); err != nil {
-			return nil, err
-		}
-
-		credentials = append(credentials, c)
-	}
-
-	return credentials, nil
-}
-
-func (s *SQLiteStore) DeleteCredentials(accessKey string) error {
-	_, err := s.db.Exec(`DELETE FROM credentials WHERE access_key = ?`, accessKey)
-	return err
-}
-
-func (s *SQLiteStore) DeleteExpiredCredentials() {
-	_, err := s.db.Exec(`DELETE FROM credentials WHERE expires_at IS NOT NULL AND DATE(expires_at) < DATE('now');`)
-	if err != nil {
-		log.Printf("Failed to delete expired credentials: %v", err)
-	} else {
-		log.Println("Expired credentials deleted.")
-	}
-}
-
-func (s *SQLiteStore) RegisterUser(username, password, accessKey, expiresAt string) error {
-	hashedPassword, err := utils.HashPassword(password)
-	if err != nil {
-		return err
-	}
-	token := utils.GenerateRandomKey(32)
-	hashedToken, err := utils.HashPassword(token)
-	if err != nil {
-		return err
-	}
-
-	var expiresAtValue interface{}
-	if expiresAt == "" {
-		expiresAtValue = nil
-	} else {
-		expiresAtValue = expiresAt
-	}
-
-	_, err = s.db.Exec(`INSERT INTO users (username, password, token, access_key, expires_at) VALUES (?, ?, ?, ?, ?);`, username, hashedPassword, hashedToken, accessKey, expiresAtValue)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (s *SQLiteStore) GetUser(username string) (models.User, error) {
-	row := s.db.QueryRow(`SELECT id, username, password, token, expires_at FROM users WHERE username = ?`, username)
-
-	var u models.User
-	if err := row.Scan(&u.Id, &u.Username, &u.Password, &u.Token, &u.ExpiresAt); err != nil {
-		return models.User{}, err
-	}
-
-	return u, nil
-}
-
-func (s *SQLiteStore) GetUserById(id int64) (models.User, error) {
-	row := s.db.QueryRow(`SELECT id, username, password, token, access_key, expires_at FROM users WHERE id = ?`, id)
-
-	var u models.User
-	if err := row.Scan(&u.Id, &u.Username, &u.Password, &u.Token, &u.AccessKey, &u.ExpiresAt); err != nil {
-		return models.User{}, err
-	}
-
-	return u, nil
-}
-
-type LoginResponse struct {
-	Id    int64
-	Token string
-}
-
-func (s *SQLiteStore) LoginUser(username, password string) (LoginResponse, error) {
-	user, err := s.GetUser(username)
-	if err != nil {
-		return LoginResponse{}, err
-	}
-	equal := utils.CheckPasswordHash(password, user.Password)
-
-	if !equal {
-		return LoginResponse{}, errors.New("invalid password")
-	}
-
-	token := utils.GenerateRandomKey(32)
-	hashedToken, err := utils.HashPassword(token)
-	if err != nil {
-		return LoginResponse{}, err
-	}
-
-	_, err = s.db.Exec(`UPDATE users SET token = ? WHERE username = ?`, hashedToken, username)
-	if err != nil {
-		return LoginResponse{}, err
-	}
-
-	return LoginResponse{Id: user.Id, Token: token}, nil
-}
-
-func (s *SQLiteStore) AuthenticateUser(id int64, token string) error {
-	user, err := s.GetUserById(id)
-	if err != nil {
-		return err
-	}
-
-	equal := utils.CheckPasswordHash(token, user.Token)
-
-	if !equal {
-		return errors.New("invalid token")
-	}
-
-	return nil
-}
-
-func (s *SQLiteStore) ListUsers() ([]models.User, error) {
-	rows, err := s.db.Query(`SELECT id, username, expires_at, created_at FROM users`)
-	if err != nil {
-		return nil, err
-	}
-
-	var users []models.User
-	for rows.Next() {
-		var u models.User
-		if err := rows.Scan(&u.Id, &u.Username, &u.ExpiresAt, &u.CreatedAt); err != nil {
-			return nil, err
-		}
-
-		users = append(users, u)
-	}
-
-	err = rows.Close()
-	if err != nil {
-		return nil, err
-	}
-
-	return users, nil
-}
-
-func (s *SQLiteStore) DeleteUser(id int64) error {
-	user, err := s.GetUserById(id)
-	if err != nil {
-		return err
-	}
-
-	err = s.DeleteCredentials(user.AccessKey)
-	if err != nil {
-		return err
-	}
-
-	_, err = s.db.Exec(`DELETE FROM users WHERE id = ?`, id)
-	return err
-}
-
-func (s *SQLiteStore) DeleteExpiredUsers() {
-	_, err := s.db.Exec(`DELETE FROM users WHERE expires_at IS NOT NULL AND DATE(expires_at) < DATE('now');`)
-	if err != nil {
-		log.Printf("Failed to delete expired users: %v", err)
-	} else {
-		log.Println("Expired users deleted.")
-	}
-}
